以下の内容を変更 - 重複コードの調整 - 機能の統合 - 保守性の向上（Gemini code review より） - ロジックの統一 - 拡張性（ページの簡略化により可能）
